Fixes from PR review.  Breaks full script test into it's own file.  Adds docs to maintest package.
