+++ .gitignore
@@ -1,6 +1,3 @@
-# Project-specific files
-bin/bpc.old
-
 # Compiled class file
 *.class
