+++ bin/bpc
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 
 #
-# Generated from bin/bpc.bps on 2024-04-30T11:15:54.123578675-07:00[America/Los_Angeles] (timestamp 1714500954123)
+# Generated from bin/bpc.bps on 2024-04-30T12:55:55.089387338-07:00[America/Los_Angeles] (timestamp 1714506955089)
 #
 # strict mode header
 set -eEuo pipefail -o posix
@@ -53,7 +53,7 @@ if [ -z "${BASHPILE_HOME+default}" ]; then
   declare bpWhere
   bpWhere=$(whereis bashpile | cut -d ' ' -f 2)
   # reassign statement, Bashpile line 34
-  bpHome=$(dirname $bpWhere)
+  bpHome=$(dirname "$bpWhere")
 else
   # assign statement, Bashpile line 36
   declare BASHPILE_HOME
