+++ docs/index.html
@@ -617,10 +617,22 @@ a { text-decoration: none; }
 <body>
 <main id="bashpile">
 <h1 class="title">Bashpile</h1>
-<p class="subtitle" id="install">Install</p>
 
-<p>Brew install
-set BASHPILE_HOME</p>
+<p>Old docs at GitHub wiki.</p>
+<section id="install">
+<h2>Install</h2>
+<ol class="arabic simple">
+<li><p>Install Homebrew if needed.</p></li>
+<li><p>Run <cite>brew install --HEAD michael-amiethyst/bashpile/bashpile</cite>.</p></li>
+</ol>
+</section>
+<section id="contributing">
+<h2>Contributing</h2>
+<ol class="arabic simple">
+<li><p>Pull code.</p></li>
+<li><p>Set BASHPILE_HOME to base of git repo.</p></li>
+</ol>
+</section>
 </main>
 </body>
 </html>
