+++ src/test/java/com/bashpile/maintests/package-info.java
@@ -1,0 +1,9 @@
+/**
+ * BashpileTest is Super to the rest of the tests.
+ * Tests are ordered by increasing complexity:
+ * LexerBashpileTest, ExpressionBashpileTest, StatementBashpileTest, FunctionBashpileTest, ShellStringBashpileTest,
+ * ConditionalsBashpileTest, ListTest, LoopsTest and FullScriptsTest
+ *
+ * TODO rename ListTest and LoopsTest
+ */
+package com.bashpile.maintests;
