+++ src/test/resources/scripts/my_ls.bash
